requirements.txt
amqp==2.1.4
ansible==2.2.1.0
anyjson==0.3.3
appdirs==1.4.0
Babel==0.9.6
backports.ssl-match-hostname==3.4.0.2
billiard==3.5.0.2
celery==4.0.2
cffi==1.9.1
configobj==4.7.2
decorator==3.4.0
Django==1.10.5
djangorestframework==3.5.4
ecdsa==0.13
enum34==1.1.6
httplib2==0.7.7
idna==2.2
iniparse==0.4
ipaddress==1.0.18
Jinja2==2.7.2
kombu==4.0.2
MarkupSafe==0.11
packaging==16.8
paramiko==1.16.1
pbr==1.10.0
perf==0.1
pyasn1==0.2.2
pycharm===debug
pycparser==2.17
pycrypto==2.6.1
pycurl==7.19.0
pydevd==1.0.0
pygobject==3.14.0
pygpgme==0.3
pyliblzma==0.5.3
pyparsing==2.1.10
python-keyczar==0.71rc0
pytz==2016.10
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
redis==2.10.5
six==1.10.0
slip==0.4.0
slip.dbus==0.4.0
stevedore==1.20.0
urlgrabber==3.10
vine==1.1.3
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
yum-metadata-parser==1.1.4
